ZadHard/python-service/python-grpc-client/requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
protobuf==5.27.0